# [Go] pack Pos field of Table in slice