--- a/go/table.go
+++ b/go/table.go
@@ -2,17 +2,37 @@ package flatbuffers
 
 // Table wraps a byte slice and provides read access to its data.
 //
-// The variable `Pos` indicates the root of the FlatBuffers object therein.
+// The byte slice `_buf` is not quite as it appears: We need to track a position
+// within the buffer (the root of the object that this Table represents), in
+// addition to the bounds of the buffer itself. Instead of using another field,
+// we put the position in the len of the slice, and the actual length in the
+// cap. When re-slicing it, `_buf` can be used as-is, but to avoid confusing
+// clients, we provide the Bytes() method to get the whole buffer.
 type Table struct {
-	Bytes []byte
-	Pos   UOffsetT // Always < 1<<31.
+	_buf []byte
+}
+
+// Init sets up the Table to refer to the object at offset `pos` in the buffer `buf`.
+// `pos` must be < 1<<31.
+func (t *Table) Init(buf []byte, pos UOffsetT) {
+	t._buf = buf[:pos:len(buf)]
+}
+
+// Bytes returns the full buffer that this Table was initialized with.
+func (t Table) Bytes() []byte {
+	return t._buf[:cap(t._buf)]
+}
+
+// Pos returns the position of the object that this Table represents within the buffer.
+func (t Table) Pos() UOffsetT {
+	return UOffsetT(len(t._buf))
 }
 
 // Offset provides access into the Table's vtable.
 //
 // Fields which are deprecated are ignored by checking against the vtable's length.
 func (t *Table) Offset(vtableOffset VOffsetT) VOffsetT {
-	vtable := UOffsetT(SOffsetT(t.Pos) - t.GetSOffsetT(t.Pos))
+	vtable := UOffsetT(SOffsetT(t.Pos()) - t.GetSOffsetT(t.Pos()))
 	if vtableOffset < t.GetVOffsetT(vtable) {
 		return t.GetVOffsetT(vtable + UOffsetT(vtableOffset))
 	}
@@ -21,7 +41,7 @@ func (t *Table) Offset(vtableOffset VOffsetT) VOffsetT {
 
 // Indirect retrieves the relative offset stored at `offset`.
 func (t *Table) Indirect(off UOffsetT) UOffsetT {
-	return off + GetUOffsetT(t.Bytes[off:])
+	return off + GetUOffsetT(t._buf[off:cap(t._buf)])
 }
 
 // String gets a string from data stored inside the flatbuffer.
@@ -32,25 +52,25 @@ func (t *Table) String(off UOffsetT) string {
 
 // ByteVector gets a byte slice from data stored inside the flatbuffer.
 func (t *Table) ByteVector(off UOffsetT) []byte {
-	off += GetUOffsetT(t.Bytes[off:])
+	off += GetUOffsetT(t._buf[off:cap(t._buf)])
 	start := off + UOffsetT(SizeUOffsetT)
-	length := GetUOffsetT(t.Bytes[off:])
-	return t.Bytes[start : start+length]
+	length := GetUOffsetT(t._buf[off:cap(t._buf)])
+	return t._buf[start : start+length]
 }
 
 // VectorLen retrieves the length of the vector whose offset is stored at
 // "off" in this object.
 func (t *Table) VectorLen(off UOffsetT) int {
-	off += t.Pos
-	off += GetUOffsetT(t.Bytes[off:])
-	return int(GetUOffsetT(t.Bytes[off:]))
+	off += t.Pos()
+	off += GetUOffsetT(t._buf[off:cap(t._buf)])
+	return int(GetUOffsetT(t._buf[off:cap(t._buf)]))
 }
 
 // Vector retrieves the start of data of the vector whose offset is stored
 // at "off" in this object.
 func (t *Table) Vector(off UOffsetT) UOffsetT {
-	off += t.Pos
-	x := off + GetUOffsetT(t.Bytes[off:])
+	off += t.Pos()
+	x := off + GetUOffsetT(t._buf[off:cap(t._buf)])
 	// data starts after metadata containing the vector length
 	x += UOffsetT(SizeUOffsetT)
 	return x
@@ -59,84 +79,84 @@ func (t *Table) Vector(off UOffsetT) UOffsetT {
 // Union initializes any Table-derived type to point to the union at the given
 // offset.
 func (t *Table) Union(t2 *Table, off UOffsetT) {
-	off += t.Pos
-	t2.Pos = off + t.GetUOffsetT(off)
-	t2.Bytes = t.Bytes
+	off += t.Pos()
+	pos := off + t.GetUOffsetT(off)
+	t2.Init(t.Bytes(), pos)
 }
 
 // GetBool retrieves a bool at the given offset.
 func (t *Table) GetBool(off UOffsetT) bool {
-	return GetBool(t.Bytes[off:])
+	return GetBool(t._buf[off:cap(t._buf)])
 }
 
 // GetByte retrieves a byte at the given offset.
 func (t *Table) GetByte(off UOffsetT) byte {
-	return GetByte(t.Bytes[off:])
+	return GetByte(t._buf[off:cap(t._buf)])
 }
 
 // GetUint8 retrieves a uint8 at the given offset.
 func (t *Table) GetUint8(off UOffsetT) uint8 {
-	return GetUint8(t.Bytes[off:])
+	return GetUint8(t._buf[off:cap(t._buf)])
 }
 
 // GetUint16 retrieves a uint16 at the given offset.
 func (t *Table) GetUint16(off UOffsetT) uint16 {
-	return GetUint16(t.Bytes[off:])
+	return GetUint16(t._buf[off:cap(t._buf)])
 }
 
 // GetUint32 retrieves a uint32 at the given offset.
 func (t *Table) GetUint32(off UOffsetT) uint32 {
-	return GetUint32(t.Bytes[off:])
+	return GetUint32(t._buf[off:cap(t._buf)])
 }
 
 // GetUint64 retrieves a uint64 at the given offset.
 func (t *Table) GetUint64(off UOffsetT) uint64 {
-	return GetUint64(t.Bytes[off:])
+	return GetUint64(t._buf[off:cap(t._buf)])
 }
 
 // GetInt8 retrieves a int8 at the given offset.
 func (t *Table) GetInt8(off UOffsetT) int8 {
-	return GetInt8(t.Bytes[off:])
+	return GetInt8(t._buf[off:cap(t._buf)])
 }
 
 // GetInt16 retrieves a int16 at the given offset.
 func (t *Table) GetInt16(off UOffsetT) int16 {
-	return GetInt16(t.Bytes[off:])
+	return GetInt16(t._buf[off:cap(t._buf)])
 }
 
 // GetInt32 retrieves a int32 at the given offset.
 func (t *Table) GetInt32(off UOffsetT) int32 {
-	return GetInt32(t.Bytes[off:])
+	return GetInt32(t._buf[off:cap(t._buf)])
 }
 
 // GetInt64 retrieves a int64 at the given offset.
 func (t *Table) GetInt64(off UOffsetT) int64 {
-	return GetInt64(t.Bytes[off:])
+	return GetInt64(t._buf[off:cap(t._buf)])
 }
 
 // GetFloat32 retrieves a float32 at the given offset.
 func (t *Table) GetFloat32(off UOffsetT) float32 {
-	return GetFloat32(t.Bytes[off:])
+	return GetFloat32(t._buf[off:cap(t._buf)])
 }
 
 // GetFloat64 retrieves a float64 at the given offset.
 func (t *Table) GetFloat64(off UOffsetT) float64 {
-	return GetFloat64(t.Bytes[off:])
+	return GetFloat64(t._buf[off:cap(t._buf)])
 }
 
 // GetUOffsetT retrieves a UOffsetT at the given offset.
 func (t *Table) GetUOffsetT(off UOffsetT) UOffsetT {
-	return GetUOffsetT(t.Bytes[off:])
+	return GetUOffsetT(t._buf[off:cap(t._buf)])
 }
 
 // GetVOffsetT retrieves a VOffsetT at the given offset.
 func (t *Table) GetVOffsetT(off UOffsetT) VOffsetT {
-	return GetVOffsetT(t.Bytes[off:])
+	return GetVOffsetT(t._buf[off:cap(t._buf)])
 }
 
 // GetSOffsetT retrieves a SOffsetT at the given offset.
 func (t *Table) GetSOffsetT(off UOffsetT) SOffsetT {
-	return GetSOffsetT(t.Bytes[off:])
+	return GetSOffsetT(t._buf[off:cap(t._buf)])
 }
 
 // GetBoolSlot retrieves the bool that the given vtable location
@@ -148,7 +168,7 @@ func (t *Table) GetBoolSlot(slot VOffsetT, d bool) bool {
 		return d
 	}
 
-	return t.GetBool(t.Pos + UOffsetT(off))
+	return t.GetBool(t.Pos() + UOffsetT(off))
 }
 
 // GetByteSlot retrieves the byte that the given vtable location
@@ -160,7 +180,7 @@ func (t *Table) GetByteSlot(slot VOffsetT, d byte) byte {
 		return d
 	}
 
-	return t.GetByte(t.Pos + UOffsetT(off))
+	return t.GetByte(t.Pos() + UOffsetT(off))
 }
 
 // GetInt8Slot retrieves the int8 that the given vtable location
@@ -172,7 +192,7 @@ func (t *Table) GetInt8Slot(slot VOffsetT, d int8) int8 {
 		return d
 	}
 
-	return t.GetInt8(t.Pos + UOffsetT(off))
+	return t.GetInt8(t.Pos() + UOffsetT(off))
 }
 
 // GetUint8Slot retrieves the uint8 that the given vtable location
@@ -184,7 +204,7 @@ func (t *Table) GetUint8Slot(slot VOffsetT, d uint8) uint8 {
 		return d
 	}
 
-	return t.GetUint8(t.Pos + UOffsetT(off))
+	return t.GetUint8(t.Pos() + UOffsetT(off))
 }
 
 // GetInt16Slot retrieves the int16 that the given vtable location
@@ -196,7 +216,7 @@ func (t *Table) GetInt16Slot(slot VOffsetT, d int16) int16 {
 		return d
 	}
 
-	return t.GetInt16(t.Pos + UOffsetT(off))
+	return t.GetInt16(t.Pos() + UOffsetT(off))
 }
 
 // GetUint16Slot retrieves the uint16 that the given vtable location
@@ -208,7 +228,7 @@ func (t *Table) GetUint16Slot(slot VOffsetT, d uint16) uint16 {
 		return d
 	}
 
-	return t.GetUint16(t.Pos + UOffsetT(off))
+	return t.GetUint16(t.Pos() + UOffsetT(off))
 }
 
 // GetInt32Slot retrieves the int32 that the given vtable location
@@ -220,7 +240,7 @@ func (t *Table) GetInt32Slot(slot VOffsetT, d int32) int32 {
 		return d
 	}
 
-	return t.GetInt32(t.Pos + UOffsetT(off))
+	return t.GetInt32(t.Pos() + UOffsetT(off))
 }
 
 // GetUint32Slot retrieves the uint32 that the given vtable location
@@ -232,7 +252,7 @@ func (t *Table) GetUint32Slot(slot VOffsetT, d uint32) uint32 {
 		return d
 	}
 
-	return t.GetUint32(t.Pos + UOffsetT(off))
+	return t.GetUint32(t.Pos() + UOffsetT(off))
 }
 
 // GetInt64Slot retrieves the int64 that the given vtable location
@@ -244,7 +264,7 @@ func (t *Table) GetInt64Slot(slot VOffsetT, d int64) int64 {
 		return d
 	}
 
-	return t.GetInt64(t.Pos + UOffsetT(off))
+	return t.GetInt64(t.Pos() + UOffsetT(off))
 }
 
 // GetUint64Slot retrieves the uint64 that the given vtable location
@@ -256,7 +276,7 @@ func (t *Table) GetUint64Slot(slot VOffsetT, d uint64) uint64 {
 		return d
 	}
 
-	return t.GetUint64(t.Pos + UOffsetT(off))
+	return t.GetUint64(t.Pos() + UOffsetT(off))
 }
 
 // GetFloat32Slot retrieves the float32 that the given vtable location
@@ -268,7 +288,7 @@ func (t *Table) GetFloat32Slot(slot VOffsetT, d float32) float32 {
 		return d
 	}
 
-	return t.GetFloat32(t.Pos + UOffsetT(off))
+	return t.GetFloat32(t.Pos() + UOffsetT(off))
 }
 
 // GetFloat64Slot retrieves the float64 that the given vtable location
@@ -280,7 +300,7 @@ func (t *Table) GetFloat64Slot(slot VOffsetT, d float64) float64 {
 		return d
 	}
 
-	return t.GetFloat64(t.Pos + UOffsetT(off))
+	return t.GetFloat64(t.Pos() + UOffsetT(off))
 }
 
 // GetVOffsetTSlot retrieves the VOffsetT that the given vtable location
@@ -296,98 +316,98 @@ func (t *Table) GetVOffsetTSlot(slot VOffsetT, d VOffsetT) VOffsetT {
 
 // MutateBool updates a bool at the given offset.
 func (t *Table) MutateBool(off UOffsetT, n bool) bool {
-	WriteBool(t.Bytes[off:], n)
+	WriteBool(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateByte updates a Byte at the given offset.
 func (t *Table) MutateByte(off UOffsetT, n byte) bool {
-	WriteByte(t.Bytes[off:], n)
+	WriteByte(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateUint8 updates a Uint8 at the given offset.
 func (t *Table) MutateUint8(off UOffsetT, n uint8) bool {
-	WriteUint8(t.Bytes[off:], n)
+	WriteUint8(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateUint16 updates a Uint16 at the given offset.
 func (t *Table) MutateUint16(off UOffsetT, n uint16) bool {
-	WriteUint16(t.Bytes[off:], n)
+	WriteUint16(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateUint32 updates a Uint32 at the given offset.
 func (t *Table) MutateUint32(off UOffsetT, n uint32) bool {
-	WriteUint32(t.Bytes[off:], n)
+	WriteUint32(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateUint64 updates a Uint64 at the given offset.
 func (t *Table) MutateUint64(off UOffsetT, n uint64) bool {
-	WriteUint64(t.Bytes[off:], n)
+	WriteUint64(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateInt8 updates a Int8 at the given offset.
 func (t *Table) MutateInt8(off UOffsetT, n int8) bool {
-	WriteInt8(t.Bytes[off:], n)
+	WriteInt8(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateInt16 updates a Int16 at the given offset.
 func (t *Table) MutateInt16(off UOffsetT, n int16) bool {
-	WriteInt16(t.Bytes[off:], n)
+	WriteInt16(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateInt32 updates a Int32 at the given offset.
 func (t *Table) MutateInt32(off UOffsetT, n int32) bool {
-	WriteInt32(t.Bytes[off:], n)
+	WriteInt32(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateInt64 updates a Int64 at the given offset.
 func (t *Table) MutateInt64(off UOffsetT, n int64) bool {
-	WriteInt64(t.Bytes[off:], n)
+	WriteInt64(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateFloat32 updates a Float32 at the given offset.
 func (t *Table) MutateFloat32(off UOffsetT, n float32) bool {
-	WriteFloat32(t.Bytes[off:], n)
+	WriteFloat32(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateFloat64 updates a Float64 at the given offset.
 func (t *Table) MutateFloat64(off UOffsetT, n float64) bool {
-	WriteFloat64(t.Bytes[off:], n)
+	WriteFloat64(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateUOffsetT updates a UOffsetT at the given offset.
 func (t *Table) MutateUOffsetT(off UOffsetT, n UOffsetT) bool {
-	WriteUOffsetT(t.Bytes[off:], n)
+	WriteUOffsetT(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateVOffsetT updates a VOffsetT at the given offset.
 func (t *Table) MutateVOffsetT(off UOffsetT, n VOffsetT) bool {
-	WriteVOffsetT(t.Bytes[off:], n)
+	WriteVOffsetT(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateSOffsetT updates a SOffsetT at the given offset.
 func (t *Table) MutateSOffsetT(off UOffsetT, n SOffsetT) bool {
-	WriteSOffsetT(t.Bytes[off:], n)
+	WriteSOffsetT(t._buf[off:cap(t._buf)], n)
 	return true
 }
 
 // MutateBoolSlot updates the bool at given vtable location
 func (t *Table) MutateBoolSlot(slot VOffsetT, n bool) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateBool(t.Pos+UOffsetT(off), n)
+		t.MutateBool(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -397,7 +417,7 @@ func (t *Table) MutateBoolSlot(slot VOffsetT, n bool) bool {
 // MutateByteSlot updates the byte at given vtable location
 func (t *Table) MutateByteSlot(slot VOffsetT, n byte) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateByte(t.Pos+UOffsetT(off), n)
+		t.MutateByte(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -407,7 +427,7 @@ func (t *Table) MutateByteSlot(slot VOffsetT, n byte) bool {
 // MutateInt8Slot updates the int8 at given vtable location
 func (t *Table) MutateInt8Slot(slot VOffsetT, n int8) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateInt8(t.Pos+UOffsetT(off), n)
+		t.MutateInt8(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -417,7 +437,7 @@ func (t *Table) MutateInt8Slot(slot VOffsetT, n int8) bool {
 // MutateUint8Slot updates the uint8 at given vtable location
 func (t *Table) MutateUint8Slot(slot VOffsetT, n uint8) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateUint8(t.Pos+UOffsetT(off), n)
+		t.MutateUint8(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -427,7 +447,7 @@ func (t *Table) MutateUint8Slot(slot VOffsetT, n uint8) bool {
 // MutateInt16Slot updates the int16 at given vtable location
 func (t *Table) MutateInt16Slot(slot VOffsetT, n int16) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateInt16(t.Pos+UOffsetT(off), n)
+		t.MutateInt16(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -437,7 +457,7 @@ func (t *Table) MutateInt16Slot(slot VOffsetT, n int16) bool {
 // MutateUint16Slot updates the uint16 at given vtable location
 func (t *Table) MutateUint16Slot(slot VOffsetT, n uint16) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateUint16(t.Pos+UOffsetT(off), n)
+		t.MutateUint16(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -447,7 +467,7 @@ func (t *Table) MutateUint16Slot(slot VOffsetT, n uint16) bool {
 // MutateInt32Slot updates the int32 at given vtable location
 func (t *Table) MutateInt32Slot(slot VOffsetT, n int32) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateInt32(t.Pos+UOffsetT(off), n)
+		t.MutateInt32(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -457,7 +477,7 @@ func (t *Table) MutateInt32Slot(slot VOffsetT, n int32) bool {
 // MutateUint32Slot updates the uint32 at given vtable location
 func (t *Table) MutateUint32Slot(slot VOffsetT, n uint32) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateUint32(t.Pos+UOffsetT(off), n)
+		t.MutateUint32(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -467,7 +487,7 @@ func (t *Table) MutateUint32Slot(slot VOffsetT, n uint32) bool {
 // MutateInt64Slot updates the int64 at given vtable location
 func (t *Table) MutateInt64Slot(slot VOffsetT, n int64) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateInt64(t.Pos+UOffsetT(off), n)
+		t.MutateInt64(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -477,7 +497,7 @@ func (t *Table) MutateInt64Slot(slot VOffsetT, n int64) bool {
 // MutateUint64Slot updates the uint64 at given vtable location
 func (t *Table) MutateUint64Slot(slot VOffsetT, n uint64) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateUint64(t.Pos+UOffsetT(off), n)
+		t.MutateUint64(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -487,7 +507,7 @@ func (t *Table) MutateUint64Slot(slot VOffsetT, n uint64) bool {
 // MutateFloat32Slot updates the float32 at given vtable location
 func (t *Table) MutateFloat32Slot(slot VOffsetT, n float32) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateFloat32(t.Pos+UOffsetT(off), n)
+		t.MutateFloat32(t.Pos()+UOffsetT(off), n)
 		return true
 	}
 
@@ -497,7 +517,7 @@ func (t *Table) MutateFloat32Slot(slot VOffsetT, n float32) bool {
 // MutateFloat64Slot updates the float64 at given vtable location
 func (t *Table) MutateFloat64Slot(slot VOffsetT, n float64) bool {
 	if off := t.Offset(slot); off != 0 {
-		t.MutateFloat64(t.Pos+UOffsetT(off), n)
+		t.MutateFloat64(t.Pos()+UOffsetT(off), n)
 		return true
 	}