… tests/MyGame/Example/Monster.go
… tests/MyGame/Example/TypeAliases.go
… tests/MyGame/InParentNamespace.go
… tests/go_test.go
… tests/namespace_test/NamespaceA/NamespaceB/StructInNestedNS.go
… tests/namespace_test/NamespaceA/NamespaceB/TableInNestedNS.go
… tests/namespace_test/NamespaceA/SecondTableInA.go
… tests/namespace_test/NamespaceA/TableInFirstNS.go
… tests/namespace_test/NamespaceC/TableInC.go