This shrinks the size of Table (and therefore all FlatBuffers objects)
from 32 bytes to 24. We're not using the capacity of the given slice for
anything, since we never modify it, so we can pack the position in the
len field, and len in cap.

-----

I had this idea while thinking about some other aspects of the Go implementation. This might provide a performance improvement for programs that create many flatbuffers objects by reducing memory pressure and making it more appealing to pass objects by value where it makes sense.

Benchmark results (from my laptop, not very rigorous):
master:
```
BenchmarkParseGold-4                    20000000               281 ns/op         795.93 MB/s           0 B/op        0 allocs/op
BenchmarkParseGold-4                    20000000               285 ns/op         785.18 MB/s           0 B/op        0 allocs/op
BenchmarkParseGold-4                    20000000               281 ns/op         795.57 MB/s           0 B/op        0 allocs/op
```
this change:
```
BenchmarkParseGold-4                    20000000               279 ns/op         802.47 MB/s           0 B/op        0 allocs/op
BenchmarkParseGold-4                    20000000               280 ns/op         799.91 MB/s           0 B/op        0 allocs/op
BenchmarkParseGold-4                    20000000               282 ns/op         793.33 MB/s           0 B/op        0 allocs/op
```

As far as the public API, client code in the same package as generated flatbuffers code that accesses the `Bytes` and `Pos` fields directly will have to change to use `Bytes()`, `Pos()`, and `Init()`. Or code that uses `flatbuffers.Table` directly. I think those should be relatively rare.