@dnr 2019-03-16
Also, just to note: Since we know all our lengths and offsets fit in 32
bits, we could actually do it in 16 bytes with one pointer and two 32-bit
fields for length and position. But that would require using `unsafe` (as
far as I can tell) and would probably be harder to maintain. So this is
sort of a middle-ground.

>