@rw 2019-05-21
@dnr I think this is clever and interesting. My main questions are:

1) Is there a great way to make this maintainable for future committers?
2) Is there a speedup that justifies this? The micro-benchmark you posted in the first comment seemed to show that there is no speed change.
3) Can you quantify the memory savings for me? My understanding is that it would save one machine word in the representation of `Table`.

Overall, I'm disinclined to include this PR, because of how it could conceivably lead to bugs in the far future.

An alternative would be to create a custom SliceHeader type that we can use, instead of manipulating the `[]byte`'s. WDYT about that?