@dnr 2019-05-21
That's a fair assessment and I don't entirely disagree. I do think that on balance, it's a positive change, though. Let me respond to your points:

1 – The "tricky bits" are entirely encapsulated in `table.go`. That file is not under active development. The last change was in 2017, which just changed a comment. The last substantial change was in 2016, and there have only been five changes since the initial version in 2014.

So, I disagree with the idea that it would create a maintenance burden.

I'm always interested in simplifying, though. I tried putting the common expression `t._buf[off:cap(t._buf)]` (which is the main piece of potentially confusing code) into a function, hoping that it would be inlined effectively, but I found a performance hit. I'm happy to try it with a newer version of the compiler or try digging into why that didn't work as well as I hoped.

2 and 3 – I think the speed and memory benefits are connected, so let me talk about them together:

Yes, we're talking about one machine word per `Table` (which of course is the sole field of all FB table objects, so we're talking about all of them).

The obvious question is how many `Table` objects there are in a typical application. I can only talk about code I've seen, but my guess is: if the code passes around pointers to FB objects, then not that many, but if it passes FB objects as values (which is generally the encouraged style in Go and saves an indirection in this case), then potentially a lot. And in that case they'd be using stack space too, which is important to conserve to avoid triggering stack copies.

So my guess at the overall effects of this change by itself would be:

If you're using pointers to FB objects: nothing much (unless you're doing something unusual).

If you're using FB objects as values: your stack would get a little smaller and you'd see a small performance benefit due to that increased locality, plus maybe a larger one if the stack usage of a critical goroutine happened to be near a copying threshold and it now avoided copying.

But of course, someone currently using pointers might decide to switch to values after this change, which would trade a little more stack space for less indirection, which might be a win.

Overall, I'd expect the effect to be small. This isn't a huge win, it's just a small optimization with the goal of avoiding a "death by a thousand cuts" situation where lots of small but avoidable losses add up to something significant.

For most other projects, I don't think this would be that important, but people use flatbuffers specifically because they do care about the cost of small allocations and copies, so I think most users would appreciate this change, even if the benefit is small.

If you'd like to see more realistic performance data, I can try building a benchmark out of an application that I'm using this in. I'm curious too, so I'll probably do this eventually, but it might take a while since there are no benchmarks of those parts currently.

Regarding your comment on a custom SliceHeader, I'm not sure exactly what you mean. Do you mean something using `unsafe`? That can work even better (we could get it down to two words), but… it's using `unsafe`, which is generally avoided without a really compelling reason. I'll keep thinking along these lines, though. There might be a way to do it with `unsafe` but a very limited use of it.